Imported all battle code. Doesn't work
I have yet to convert every command to xLIB commands and stuff like that, but most of the work is there. Menu code and stuff works perfectly fine, and I believe that it passes attacks correctly and may display them correctly. I'm too tired to continue working on this today lol